moving the rest of this refactor pass to the right files
